# clients/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mobility Management Agent API",
    "summary": "Governed, evidence-backed airport-access mobility concierge",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/trips/candidates": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Parse Trip Candidate",
        "operationId": "parse_trip_candidate_api_v1_trips_candidates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripCandidate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/candidates/image": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Parse Trip Image",
        "operationId": "parse_trip_image_api_v1_trips_candidates_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_trip_image_api_v1_trips_candidates_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripCandidate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/decisions/preview": {
      "post": {
        "tags": [
          "decisions"
        ],
        "summary": "Preview Decision",
        "operationId": "preview_decision_api_v1_decisions_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assistant/questions": {
      "post": {
        "tags": [
          "assistant"
        ],
        "summary": "Answer Question",
        "operationId": "answer_question_api_v1_assistant_questions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantAnswer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reminders/preview": {
      "post": {
        "tags": [
          "reminders"
        ],
        "summary": "Preview Reminder",
        "operationId": "preview_reminder_api_v1_reminders_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/action-proposals": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Propose Action",
        "operationId": "propose_action_api_v1_action_proposals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionProposal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionProposal": {
        "properties": {
          "proposal_id": {
            "type": "string",
            "title": "Proposal Id"
          },
          "action_type": {
            "type": "string",
            "enum": [
              "open_map",
              "open_ride_hailing"
            ],
            "title": "Action Type"
          },
          "status": {
            "type": "string",
            "const": "awaiting_user_confirmation",
            "title": "Status"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "parameters_preview": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Parameters Preview"
          },
          "deep_link": {
            "type": "string",
            "title": "Deep Link"
          },
          "fallback_url": {
            "type": "string",
            "title": "Fallback Url"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "requires_payment": {
            "type": "boolean",
            "const": false,
            "title": "Requires Payment",
            "default": false
          },
          "automatic_booking": {
            "type": "boolean",
            "const": false,
            "title": "Automatic Booking",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "proposal_id",
          "action_type",
          "status",
          "label",
          "parameters_preview",
          "deep_link",
          "fallback_url",
          "expires_at",
          "idempotency_key"
        ],
        "title": "ActionProposal"
      },
      "ActionProposalRequest": {
        "properties": {
          "trip": {
            "$ref": "#/components/schemas/TripInput"
          },
          "decision": {
            "$ref": "#/components/schemas/DepartureDecision"
          },
          "action_type": {
            "type": "string",
            "enum": [
              "open_map",
              "open_ride_hailing"
            ],
            "title": "Action Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "trip",
          "decision",
          "action_type"
        ],
        "title": "ActionProposalRequest"
      },
...
      "AssistantAnswer": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "cited_evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cited Evidence Ids"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "evidence_only": {
            "type": "boolean",
            "const": true,
            "title": "Evidence Only",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "answer",
          "cited_evidence_ids",
          "provider"
        ],
        "title": "AssistantAnswer"
      },
      "AssistantQuestionRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 300,
            "minLength": 2,
            "title": "Question"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionResponse"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question",
          "decision"
        ],
        "title": "AssistantQuestionRequest"
      },
      "Body_parse_trip_image_api_v1_trips_candidates_image_post": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          },
          "departure_place": {
            "type": "string",
            "title": "Departure Place",
            "default": "待确认出发地"
          },
          "checked_baggage": {
            "type": "boolean",
            "title": "Checked Baggage",
            "default": false
          },
          "risk_profile": {
            "$ref": "#/components/schemas/RiskProfile",
            "default": "cautious"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_parse_trip_image_api_v1_trips_candidates_image_post"
      },
...
      "Coordinates": {
        "properties": {
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "longitude",
          "latitude"
        ],
        "title": "Coordinates"
      },
...
      "DecisionResponse": {
        "properties": {
          "trip": {
            "$ref": "#/components/schemas/TripInput"
          },
          "context": {
            "$ref": "#/components/schemas/JourneyContext"
          },
          "decision": {
            "$ref": "#/components/schemas/DepartureDecision"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Evidence"
          },
          "assistant_summary": {
            "type": "string",
            "title": "Assistant Summary"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeBoundary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "trip",
          "context",
          "decision",
          "evidence",
          "assistant_summary",
          "verified",
          "runtime"
        ],
        "title": "DecisionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JourneyContext": {
        "properties": {
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "data_scope": {
            "type": "string",
            "enum": [
              "synthetic",
              "mixed",
              "live"
            ],
            "title": "Data Scope"
          },
          "flight": {
            "$ref": "#/components/schemas/FlightSnapshot"
          },
          "airport": {
            "$ref": "#/components/schemas/AirportProcessSnapshot"
          },
          "route": {
            "$ref": "#/components/schemas/RouteSnapshot"
          },
          "weather": {
            "$ref": "#/components/schemas/WeatherSnapshot"
          },
          "disruptions": {
            "items": {
              "$ref": "#/components/schemas/DisruptionSignal"
            },
            "type": "array",
            "title": "Disruptions"
          },
          "missing_sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Sources"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "observed_at",
          "data_scope",
          "flight",
          "airport",
          "route",
          "weather",
          "disruptions",
          "missing_sources",
          "warnings"
        ],
        "title": "JourneyContext"
      },
...
      "ReminderPreview": {
        "properties": {
          "reminder_id": {
            "type": "string",
            "title": "Reminder Id"
          },
          "remind_at": {
            "type": "string",
            "format": "date-time",
            "title": "Remind At"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "string",
            "const": "preview",
            "title": "Status"
          },
          "channel_options": {
            "items": {
              "type": "string",
              "enum": [
                "calendar",
                "web_notification",
                "wechat_subscription"
              ]
            },
            "type": "array",
            "title": "Channel Options"
          },
          "requires_user_consent": {
            "type": "boolean",
            "const": true,
            "title": "Requires User Consent",
            "default": true
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "calendar_ics": {
            "type": "string",
            "title": "Calendar Ics"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reminder_id",
          "remind_at",
          "title",
          "message",
          "status",
          "channel_options",
          "idempotency_key",
          "calendar_ics"
        ],
        "title": "ReminderPreview"
      },
      "ReminderPreviewRequest": {
        "properties": {
          "trip": {
            "$ref": "#/components/schemas/TripInput"
          },
          "decision": {
            "$ref": "#/components/schemas/DepartureDecision"
          },
          "lead_hours": {
            "type": "integer",
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Lead Hours",
            "default": 24
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "trip",
          "decision"
        ],
        "title": "ReminderPreviewRequest"
      },
      "RiskProfile": {
        "type": "string",
        "enum": [
          "standard",
          "cautious",
          "very_cautious"
        ],
        "title": "RiskProfile"
      },
...
      "TripCandidate": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "source_type": {
            "$ref": "#/components/schemas/TripSourceType"
          },
          "source_summary": {
            "type": "string",
            "title": "Source Summary"
          },
          "flight_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flight Number"
          },
          "departure_airport": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departure Airport"
          },
          "destination_airport": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Airport"
          },
          "terminal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal"
          },
          "scheduled_departure": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Departure"
          },
          "departure_place": {
            "type": "string",
            "title": "Departure Place"
          },
          "checked_baggage": {
            "type": "boolean",
            "title": "Checked Baggage"
          },
          "risk_profile": {
            "$ref": "#/components/schemas/RiskProfile"
          },
          "field_confidence": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Field Confidence"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "redactions_applied": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Redactions Applied"
          },
          "needs_user_confirmation": {
            "type": "boolean",
            "const": true,
            "title": "Needs User Confirmation",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "candidate_id",
          "source_type",
          "source_summary",
          "departure_place",
          "checked_baggage",
          "risk_profile",
          "field_confidence",
          "missing_fields",
          "warnings",
          "redactions_applied"
        ],
        "title": "TripCandidate"
      },
      "TripInput": {
        "properties": {
          "flight_number": {
            "type": "string",
            "pattern": "^[A-Z0-9]{2}\\d{3,4}$",
            "title": "Flight Number"
          },
          "departure_airport": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "pattern": "^[A-Z]{3}$",
            "title": "Departure Airport"
          },
          "destination_airport": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 3,
                "minLength": 3,
                "pattern": "^[A-Z]{3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Airport"
          },
          "terminal": {
            "type": "string",
            "maxLength": 12,
            "minLength": 1,
            "title": "Terminal",
            "default": "待确认"
          },
          "scheduled_departure": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled Departure"
          },
          "departure_place": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "title": "Departure Place"
          },
          "departure_coordinates": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Coordinates"
              },
              {
                "type": "null"
              }
            ]
          },
          "checked_baggage": {
            "type": "boolean",
            "title": "Checked Baggage",
            "default": false
          },
          "accessibility_assistance": {
            "type": "boolean",
            "title": "Accessibility Assistance",
            "default": false
          },
          "risk_profile": {
            "$ref": "#/components/schemas/RiskProfile",
            "default": "cautious"
          },
          "live_data_consent": {
            "type": "boolean",
            "title": "Live Data Consent",
            "default": false
          },
          "model_egress_consent": {
            "type": "boolean",
            "title": "Model Egress Consent",
            "default": false
          },
          "user_disruption_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "User Disruption Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "flight_number",
          "departure_airport",
          "scheduled_departure",
          "departure_place"
        ],
        "title": "TripInput"
      },
      "TripParseRequest": {
        "properties": {
          "source_type": {
            "type": "string",
            "enum": [
              "text",
              "ics"
            ],
            "title": "Source Type"
          },
          "content": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 3,
            "title": "Content"
          },
          "departure_place": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "title": "Departure Place",
            "default": "待确认出发地"
          },
          "checked_baggage": {
            "type": "boolean",
            "title": "Checked Baggage",
            "default": false
          },
          "risk_profile": {
            "$ref": "#/components/schemas/RiskProfile",
            "default": "cautious"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_type",
          "content"
        ],
        "title": "TripParseRequest"
      },
      "TripSourceType": {
        "type": "string",
        "enum": [
          "manual",
          "text",
          "ics",
          "image"
        ],
        "title": "TripSourceType"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}